# docs-standalone/openapi.yaml
openapi: 3.0.3
info:
  title: Examify API Documentation
  version: 1.0.0
  description: >-
    **Examify** is a production-grade, self-hosted AI Exam Generation &
    Assessment Platform built on Next.js 15 (App Router), MariaDB (Prisma ORM),
    Redis, and multi-model AI adapters (Google Gemini, OpenAI, Anthropic Claude,
    and Self-Hosted/vLLM/Ollama).


    ### Core Security & Platform Highlights:

    - **Zero Cloudflare Dependencies**: Self-hosted Redis Sliding Window rate
    limiting and MaxMind GeoIP resolution.

    - **Strict Role-Based Access Control**: `ADMIN` and `TEACHER` roles with
    token blacklisting in Redis.

    - **Two-Factor Authentication (2FA)**: Built-in TOTP (Google Authenticator)
    with QR code setup and challenge verification.

    - **Anti-Prompt Injection Guard**: Advanced multi-tier sanitization and
    jailbreak defense for teacher custom instructions.

    - **Permanent Initialization Lock**: One-time initial setup endpoint locked
    permanently once root admin is created.

    - **Comprehensive Audit Trail**: Immutable system logs recording security
    events, authentication attempts, AI synthesis, and administrative updates.
  contact:
    name: Examify Engineering Team
    email: admin@examify.local
    url: http://localhost:3000
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: Examify API v1 Base Path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Provide your JWT bearer token in the Authorization header (e.g.
        `Bearer eyJhbGciOi...`).
  schemas:
    StandardResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
        data:
          type: object
...
    TeacherItem:
      type: object
      properties:
        id:
          type: string
          example: GV-001
        name:
          type: string
          example: ThS. Nguyễn Văn An
        email:
          type: string
          example: an.nguyen@spt.edu.vn
        role:
          type: string
          example: Tổ trưởng chuyên môn
        department:
          type: string
          example: Toán - Tin học
        twoFactorEnabled:
          type: boolean
          example: true
        status:
          type: string
          enum:
            - active
            - suspended
          example: active
        tokensUsed:
          type: integer
          example: 142500
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /admin/teachers:
    get:
      tags:
        - Admin - Teachers
      summary: Danh sách tài khoản giáo viên & Thống kê sử dụng (ADMIN only)
      description: Lấy danh sách tất cả giáo viên trong trường, số lượng token AI đã
        dùng, trạng thái 2FA và trạng thái tài khoản.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Danh sách giáo viên
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeacherItem"
    post:
      tags:
        - Admin - Teachers
      summary: Thêm tài khoản giáo viên mới trực tiếp (ADMIN only)
      description: Tạo tài khoản giáo viên với mật khẩu khởi tạo, tự động gán tổ bộ
        môn và quyền hạn giảng dạy.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  example: Cô Nguyễn Thị Hoa
                email:
                  type: string
                  format: email
                  example: hoa.nguyen@spt.edu.vn
                role:
                  type: string
                  default: Giáo viên
                  example: Giáo viên chính khóa
                department:
                  type: string
                  default: Toán học
                  example: Khoa học Tự nhiên
                password:
                  type: string
                  minLength: 6
                  default: Examify@123
      responses:
        "201":
          description: Thêm giáo viên thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
    put:
      tags:
        - Admin - Teachers
      summary: Cập nhật trạng thái hoặc Đặt lại 2FA của giáo viên (ADMIN only)
      description: Khóa / mở khóa tài khoản giáo viên (`active` / `suspended`), hoặc
        đặt lại trạng thái 2FA nếu giáo viên mất thiết bị xác thực.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  example: GV-001
                status:
                  type: string
                  enum:
                    - active
                    - suspended
                reset2FA:
                  type: boolean
                  description: Đặt true nếu muốn vô hiệu hóa 2FA để giáo viên cài lại
      responses:
        "200":
          description: Cập nhật tài khoản giáo viên thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
  /admin/security:
    get:
      tags:
        - Admin - Security & GeoIP
      summary: Xem cấu hình bảo mật, Rate Limit & Chặn quốc gia GeoIP (ADMIN only)
      description: Lấy cấu hình cửa sổ giới hạn lượt gọi (Sliding Window Rate Limiter)
        và danh sách quốc gia được phép / bị chặn truy cập.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Cấu hình bảo mật hệ thống
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      rateLimiting:
                        type: object
                        properties:
                          windowSeconds:
                            type: integer
                            example: 60
                          authMaxRequests:
                            type: integer
                            example: 20
                          apiMaxRequests:
                            type: integer
                            example: 120
                      geoBlocking:
                        type: object
                        properties:
                          enabled:
                            type: boolean
                            example: true
                          defaultAction:
                            type: string
                            enum:
                              - ALLOW
                              - BLOCK
                            example: ALLOW
                          allowedCountries:
                            type: array
                            items:
                              type: string
                            example:
                              - VN
                              - US
                              - JP
                          blockedCountries:
                            type: array
                            items:
                              type: string
    post:
      tags:
        - Admin - Security & GeoIP
      summary: Cập nhật chính sách Rate Limiting & GeoIP Blocking (ADMIN only)
      description: Điều chỉnh giới hạn số yêu cầu mỗi phút và chính sách chặn địa lý
        (GeoIP Blocking) tự lưu trữ.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rateLimiting:
                  type: object
                  properties:
                    windowSeconds:
                      type: integer
                      default: 60
                    authMaxRequests:
                      type: integer
                      default: 20
                    apiMaxRequests:
                      type: integer
                      default: 120
                geoBlocking:
                  type: object
                  properties:
                    enabled:
                      type: boolean
                    defaultAction:
                      type: string
                      enum:
                        - ALLOW
                        - BLOCK
                    allowedCountries:
                      type: array
                      items:
                        type: string
                    blockedCountries:
                      type: array
                      items:
                        type: string
      responses:
        "200":
          description: Lưu cấu hình bảo mật thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
    put:
      tags:
        - Admin - Security & GeoIP
      summary: Mô phỏng kiểm tra quyền truy cập của một địa chỉ IP (ADMIN only)
      description: Công cụ chẩn đoán nhanh giúp kiểm tra một IP cụ thể (hoặc mã quốc
        gia) sẽ bị chặn hay được phép truy cập theo chính sách hiện hành.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ip:
                  type: string
                  example: 14.161.22.45
                country:
                  type: string
                  example: VN
      responses:
        "200":
          description: Kết quả mô phỏng đánh giá truy cập
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      testIp:
                        type: string
                        example: 14.161.22.45
                      testCountry:
                        type: string
                        example: VN
                      isAllowed:
                        type: boolean
                        example: true
                      reason:
                        type: string
                        example: Quốc gia VN nằm trong danh sách cho phép
...